refactor: add validation return in delete playlist category use case

+++ libs/domain/src/lib/use-case/playlist/category/delete-playlist-category.ts
@@ -34,12 +34,23 @@ export class DeletePlaylistCategory
       return left(new EntityNotEmpty('Logged User ID'));
     }
 
-    await ValidationUserId(loggedUserId, this.findUserByIdRepository);
+    const userValidation = await ValidationUserId(
+      loggedUserId,
+      this.findUserByIdRepository
+    );
 
-    await ValidationPlaylistCategoryId(
+    if (userValidation.isLeft()) {
+      return left(userValidation.value);
+    }
+
+    const playlistCategoryValidation = await ValidationPlaylistCategoryId(
       id,
       this.findPlaylistCategoryByIdRepository
     );
+
+    if (playlistCategoryValidation.isLeft()) {
+      return left(playlistCategoryValidation.value);
+    }
 
     await this.deletePlaylistRepository.delete(id);
